Refactor tasks (#63)
* Methods common to multiple tasks added to abstract class Tasks

* Further refactoring should be done post-architectural decisions
