+++ src/main/java/duke/tasks/Within.java
@@ -6,8 +6,6 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 
 public class Within extends Task {
-    private DateTime endDate;
-
     /**
      * Constructor for Duke.Tasks.Within object.
      * @param description name of the task.
@@ -40,16 +38,18 @@ public class Within extends Task {
     }
 
     @Override
-    public String toString() {
-        return "[W]" + super.toString() + " (between: " + this.getStartDateString()
-                + " and " + this.getEndDateString() + ")";
-    }
-
-    private String getStartDateString() {
+    String getStartDateString() {
         return this.startDate.toString();
     }
 
-    private String getEndDateString() {
+    @Override
+    String getEndDateString() {
         return this.endDate.toString();
+    }
+
+    @Override
+    public String toString() {
+        return "[W]" + super.toString() + " (between: " + this.getStartDateString()
+                + " and " + this.getEndDateString() + ")";
     }
 }
